requirements.txt
# PDF Retypeset 依赖

# 图像处理
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# PDF处理
PyMuPDF>=1.23.0
pdf2image>=1.16.0

# 配置管理
PyYAML>=6.0
